+++ lib/experimental/editor-settings.php
@@ -39,6 +39,9 @@ function gutenberg_enable_experiments() {
 	if ( gutenberg_is_experiment_enabled( 'gutenberg-dashboard-widgets' ) ) {
 		wp_add_inline_script( 'wp-block-editor', 'window.__experimentalDashboardWidgets = true', 'before' );
 	}
+	if ( gutenberg_is_experiment_enabled( 'gutenberg-media-smart-crop' ) ) {
+		wp_add_inline_script( 'wp-block-editor', 'window.__experimentalMediaSmartCrop = true', 'before' );
+	}
 }
 
 add_action( 'admin_init', 'gutenberg_enable_experiments' );